Refactor: socket object passed as props to avoid using sockets as global variable
