+++ client/src/App.tsx
@@ -1,17 +1,19 @@
 import "./App.css";
-// import { ChangeEvent, FormEvent } from "react";
 import { useEffect } from "react";
 import { useState } from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Home from "./pages/HomePage";
-import Lobby from "./pages/Lobby";
+import Home from "./pages/HomePage/HomePage";
+import Lobby from "./pages/Lobby/Lobby";
 import socketService from "./services/socketServices";
 
 function App() {
  const [lobbyId, setLobbyId] = useState("");
  const [isLoading, setIsLoading] = useState(false);
 
- // The better way with error handling: 
+ // At the moment, the socketService class gets instantiated when the .connect() method is called.
+ // Can consider refactoring to use a custom hook or useContext() api.
+
+ // The better way with error handling:
  //  const connectSocket = async () => {
  //   try {
  //    await socketService.connect("http://localhost:3001");
@@ -23,8 +25,8 @@ function App() {
 
  const connectSocket = () => {
   socketService.connect("http://localhost:3001");
-  if(socketService){
-    setIsLoading(true);
+  if (socketService) {
+   setIsLoading(true);
   }
  };
 
@@ -39,7 +41,10 @@ function App() {
  return (
   <Router>
    <Routes>
-    <Route path="/" element={<Home setLobbyId={setLobbyId} />} />
+    <Route
+     path="/"
+     element={<Home setLobbyId={setLobbyId} socketService={socketService} />}
+    />
     <Route path={`/lobby/${lobbyId}`} element={<Lobby lobbyId={lobbyId} />} />
    </Routes>
   </Router>
