+++ client/src/services/socketServices.ts
@@ -1,9 +1,11 @@
 import { io, Socket } from "socket.io-client";
 
 // To make this more advanced, think about promises and errors
+// let instance; 
 
 class SocketService {
  private socket: Socket | null = null;
+
 
 // The more safe way with error handling: 
 //  public connect(url: string): Promise<void> {
@@ -22,18 +24,23 @@ class SocketService {
 
  public connect(url: string): void {
   this.socket = io(url);
-  console.log("Client - connected successfully")
-//   console.log(this.socket)
+  console.log("Client - connected successfully:", url)
  }
 
  public getSocket(): Socket | null {
   return this.socket;
  }
+
  public on(event: string, callback: (data: any) => void): void {
   if (this.socket) {
    this.socket.on(event, callback);
-   console.log(`On: ${event}`)
+   console.log(`On event registered: ${event}`)
   }
+ }
+
+ public off(event: string){
+  this.socket?.off(event)
+  console.log(`Off event registered: ${event}`)
  }
 
  public emit(event: string, data?: any): void {
@@ -44,7 +51,8 @@ class SocketService {
  }
 }
 
-const socketService = new SocketService();
+// const socketService = new SocketService();
+const socketService = new SocketService(); 
 export default socketService;
