+++ client/tests/pages/lobby/Lobby.test.tsx
@@ -1,4 +1,4 @@
-import Lobby from "../../../src/pages/Lobby";
+import Lobby from "../../../src/pages/Lobby/Lobby";
 import { describe, it, expect } from "vitest";
 import { render, screen } from "@testing-library/react";
 import React from "react";
